4.2.2 -  Diseño e implementación de un frecuencímetro

+++ pract-2/422-diseno-e-implementacion-de-un-frecuencimetro.ino
@@ -82,7 +82,7 @@ int contador;
 int increment;
 
 volatile int frecuencia;
-volatile int periodo;
+volatile float periodo;
 int ICR3_old;
 int ICR3_new;
